feat: split onboard state
